requirements.txt
numpy==1.23.1
opencv-contrib-python==4.4.0.46

scikit-learn~=1.1.3
joblib~=1.2.0
Pillow~=9.3.0
mediapipe~=0.9.0
pyglet~=2.0.0
playsound~=1.3.0
pyobjc~=9.0